requirements.txt
# PubMed biomedical RAG + KG + LangGraph (local-first)
biopython>=1.83
chromadb>=0.5.0
sentence-transformers>=3.0.0
torch>=2.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
networkx>=3.2
tenacity>=8.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pytest>=8.0.0
rank-bm25>=0.2.2
scikit-learn>=1.3.0
